Extract MissionHoverCard into reusable library UI
